Create gen_TA_template.py
An automated script to for every row (except header), copy the header and the row to the specific sheet which name would be "DM-<placeholder>". Meanwhile the <placeholder> value can be gain in column named "data model". e.g. Sheet "DM-Authentication"  or sheet "DM-Web" with a condition where there's a gap of 40 rows between last entry